Initial commit
This first commit holds everything I learned about the new technologies
I'm studying. There's an example of pretty much eveything:

* Transpilation
* Templates
* Testing
* Markdown code inside Jade code
* Arbitrary code inside Markdown code inside Jade code

WARNING: on Windows systems an `npm install` will likely fail due to a silly
limitation on path lengths. The only way to solve this is to check which
package failed to install, install it manually and then move it to the
intended location.